创建了登录与注册模块。增加了鉴权中间件。增加了jwt的验证。

+++ internal/pkg/jwt/jwt.go
@@ -10,11 +10,11 @@ import (
 
 type MyClaims struct {
 	Username string `json:"username"`
-	UserID   uint   `json:"user_id"`
+	UserID   uint64 `json:"user_id"`
 	jwt.RegisteredClaims
 }
 
-func GenerateToken(username string, userID uint) (string, error) {
+func GenerateToken(username string, userID uint64) (string, error) {
 	claims := MyClaims{
 		username,
 		userID,
